requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
Django==4.2.27
django-crispy-forms==2.5
django-environ==0.12.0
django-ninja==1.5.3

djangorestframework==3.16.1
dotenv==0.9.9
environ==1.0
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
